+++ src/ShokoTD++lib/AlignmentVector.cpp
@@ -9,20 +9,20 @@ AlignmentVector::AlignmentVector(unsigned int _good, unsigned int _evil) :
 
 unsigned int AlignmentVector::GetGoodScore()
 {
-	double mag = sqrtf(good_ * good_ + evil_ * evil_);
+	double mag = sqrt(static_cast<double>(good_ * good_ + evil_ * evil_));
 	double sum = good_ + evil_;
 	double good = (double)good_ / mag;
 	good = good * good;
 
-	return good * sum;
+	return static_cast<unsigned int>(good * sum);
 }
 
 unsigned int AlignmentVector::GetEvilScore()
 {
-	double mag = sqrtf(good_ * good_ + evil_ * evil_);
+	double mag = sqrt(static_cast<double>(good_ * good_ + evil_ * evil_));
 	double sum = good_ + evil_;
 	double evil = (double)evil_ / mag;
 	evil = evil * evil;
 
-	return evil * sum;
+	return static_cast<unsigned int>(evil * sum);
 }
